+++ lib/dialogs.dart
@@ -3,3 +3,5 @@ library dialogs;
 export 'src/button.dart';
 export 'src/dialogs.dart';
 export 'src/helper.dart';
+export 'src/snack_bars/fancy.dart';
+export 'src/snack_bars/overlay.dart';